+++ CHANGELOG.md
@@ -78,6 +78,7 @@ v5.4.0
 - Add a check to avoid unnecessary copy in DoOdoProjecton function (!1582)
 - Update user guide and developer guide (!1598)
 - Some various tidy-up (!1585)
+- Added support for performance tests (!1614)
 
 **NekMesh**
 - Changed CMake to set NEKTAR_USE_THREAD_SAFETY to ON when MeshGen is set to ON (!1546)